==> 0.1 Retailrocket Preprocessing - Filtered items with min 30 actions.ipynb <==
import polars as pl
import os

# load the data as a lazy dataframe
data = pl.scan_csv(r'data\downloaded datasets\retailrocket\events.csv',
                   with_column_names=lambda names: ['Time', 'UserId', 'Type', 'ItemId', 'transactionid'])

print(data.fetch(5))
print(data.collect().shape)

# remove transactionid colum
data = data.drop('transactionid')

# check for missing values
data.null_count().collect()

Create session ids. each session has a maximum lenght of 30 minutes.

# create a timestamp column 
data = (data
          .with_columns(
              pl.from_epoch("Time", time_unit="ms")
              )
          )

data.collect()

create session ids. Each session has a maximum lenght of 30 minutes.

# sort the dataset on userid and timestamp
data = data.sort(['UserId', 'Time'])

data = (data
       # create two new columns that contain boolean values if the previous row is another user or exceeds the maximum session length
       .with_columns([
           # check if the user_id is different from the previous row
           (pl.col("UserId").diff() != 0).fill_null(True).alias("char_diff"),
           # check if the time difference between the current and previous row exceeds the threshold
           (pl.col("Time").diff().dt.total_minutes().cast(pl.Int64) > 30).fill_null(True).alias("ts_diff"),
       ])

       # create a new column that determines if a new session needs to be started after this row
       .with_columns([
           (pl.col("ts_diff") | pl.col("char_diff")).alias("new_session_mark")
       ])

       # drop the columns that we don't need anymore
       .drop(["char_diff", "ts_diff"])

       # create a new column that contains the session id
       .with_columns([
           (pl.col("new_session_mark").cum_sum()).alias("SessionId")
       ])
       # drop the column that we don't need anymore
       .drop("new_session_mark")
       )
data.collect()

Filter out session ids with only one interaction and also session ids that have items that appear les than 5 times. (item cold start)

# Get session ids with more than one interaction in session
SessionIdsWithMoreThanOneInteraction = (data
                                        # get the length of each session
                                        .group_by('SessionId').len()
                                        # filter out the sessions that only contain one action
                                        .filter(pl.col('len') > 1)
                                        .select('SessionId')
                                        ).collect()

SessionIdsWithMoreThanOneInteraction

ItemIdsColdstartItems = (data
                         .group_by(['ItemId', 'Type']).len()
                         .filter(pl.col('Type') == 'view')
                         .filter(pl.col('len') < 30)
                         .select('ItemId')
                         ).collect()

print(f'ItemIdsColdstartItems: {ItemIdsColdstartItems}')

SessionIdsWithColdstartItems = (data
                                .filter(pl.col('ItemId').is_in(ItemIdsColdstartItems))
                                .select('SessionId')
                                ).unique().collect()
print(f'SessionIdsWithColdstartItems: {SessionIdsWithColdstartItems}')

# keep the session ids that have more than one interaction and do not contain coldstart items
data = (data
        .filter(pl.col('SessionId').is_in(SessionIdsWithMoreThanOneInteraction))
        # .filter(~pl.col('ItemId').is_in(ItemIdsColdstartItems))
        .filter(~pl.col('SessionId').is_in(SessionIdsWithColdstartItems))

        # sort the dataset on userid and timestamp
        .sort(['SessionId', 'Time'])
        ).collect()

display(data)
data.shape

Create a train and test set.

min_time = data['Time'].min()
max_time = data['Time'].max()

print(f"First date in the dataset: {min_time}")
print(f"Last date in the dataset: {max_time}")

# take last 7 days of data as test set
test = data.filter(pl.col('Time') > max_time - pl.duration(days=7))

test

# remove test data from the dataset
train = data.filter(pl.col('Time') <= max_time - pl.duration(days=7))

train

print(f"Sessions in train set: {train['SessionId'].n_unique()}")

Filter out items that are in the test data but not in the train data. 

ItemIdsInTest = test.select('ItemId').unique()
ItemIdsInTrain = train.select('ItemId').unique()

# create a list with the item ids that are in the test set but not in the training set
ItemIdsInTestbutNotInTrain = ItemIdsInTest.filter(~pl.col('ItemId').is_in(ItemIdsInTrain))

# remove items that are in the test set but not in the training set
test = test.filter(~pl.col('ItemId').is_in(ItemIdsInTestbutNotInTrain))

test.shape

prepare the data for the model by converting the Time back to unix time

train = train.with_columns(
    pl.col("Time").dt.epoch(time_unit='s').alias("Time")
)

test = test.with_columns(
    pl.col("Time").dt.epoch(time_unit='s').alias("Time")
)   

save the datasets

# create directory
os.makedirs('data/processed datasets/retailrocket', exist_ok=True)

train.write_csv('data/processed datasets/retailrocket/events_train - Filtered items with min 30 actions.csv',
                separator='\t')
test.write_csv('data/processed datasets/retailrocket/events_test - Filtered items with min 30 actions.csv',
               separator='\t')

summary statistics

# concat the train and test data to calculate the statistics
data = pl.concat([train, test])

data.group_by('Type').len()

data.shape

def summary_stats(df, dataset_str):
    print(f'Summary statistics for the {dataset_str} dataset: ')
    n_events = df.shape[0]
    print(f'Number of events: {n_events}')
    # calculate the number of unique items
    n_items = df.select('ItemId').n_unique()
    print(f'Number of unique items: {n_items}')
    # calculate the number of sessions
    n_sessions = df.select('SessionId').n_unique()
    print(f'Number of sessions: {n_sessions}')
    # calculate the session length
    session_length = df.group_by('SessionId').agg(pl.count('SessionId').alias('len')).select(pl.col('len').mean())[0, 0]
    print(f'Average session length: {round(session_length, 2)}\n')
    
summary_stats(data, 'whole')
summary_stats(train, 'train')
summary_stats(test, 'test')
