==> 1. Digi & Ret Evaluation VSTAN baseline - Filtered items with min 30 actions.ipynb <==
import pandas as pd
from src.recsys_baseline import VSKNN_STAN
from src.evaluation import evaluate_sessions, MRR, HitRate
from src.utils import evaluation_results_to_csv

# Item knn sesion based recommendation
This is the base model that will be used to compare the RL model with.

# Initialise the model (`VSKNN_STAN`)
Model parameters are based on research of [S. Latifi, N. Mauro and D. Jannach. 2021. Session-aware recommendation: a surprising quest for the state-of-the-art. Information Sciences](https://doi.org/10.1016/j.ins.2021.05.048) 

* Optimised parameters for Diginetica dataset: [config](https://github.com/rn5l/session-rec/blob/5dcd583cbd8d44703a5248b9a308945f24b91390/conf/save/diginetica/window/window_multiple_digi_vstan.yml)  
    k: 100  
    sample_size: 1000  
    similarity: 'vec'  
    stan:  
    lambda_spw: 4.9  
    lambda_snh: 80  
    lambda_inh: 9.8  
    vsknn:  
    lambda_ipw: 4.9  
    lambda_idf: 5  


* Optimised parameters for Retail Rocket dataset: [config](https://github.com/rn5l/session-rec/blob/5dcd583cbd8d44703a5248b9a308945f24b91390/conf/save/retailrocket/session_based/window/window_retailr_vstan.yml)  
    k: 500  
    sample_size: 1000  
    similarity: 'cosine'  
    stan:  
    lambda_spw: 7.24  
    lambda_snh: 100  
    lambda_inh: 3.62  
    vsknn:  
    lambda_ipw: 3.62  
    lambda_idf: 1  

## Diginetica

train_data = pd.read_csv(r'data\processed datasets\diginetica\interactions_train_tr.txt',
                    sep='\t',)
display(train_data.head())
display(train_data.shape)

test_data = pd.read_csv(r'data\processed datasets\diginetica\interactions_test.txt',
                    sep='\t',)

display(test_data.head())
display(test_data.shape)

model_Digi = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

model_Digi.fit(train=train_data, test=test_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5),
    HitRate(10),
    HitRate(15),
    HitRate(20)]

output = evaluate_sessions(pr=model_Digi,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='VSTAN (Baseline)', reward_func='no reward func')

## Retail Rocket

train_data = pd.read_csv(r'data\processed datasets\retailrocket\events_train - Filtered items with min 30 actions.csv',
                    sep='\t',)
display(train_data.head())
display(train_data.shape)

test_data = pd.read_csv(r'data\processed datasets\retailrocket\events_test - Filtered items with min 30 actions.csv',
                    sep='\t',)

display(test_data.head())
display(test_data.shape)

model_Ret = VSKNN_STAN(k=500,
                       sample_size=1000,
                       similarity='cosine',  
                       #stan:  
                       lambda_spw=7.24,  
                       lambda_snh=100,  
                       lambda_inh=3.62,  
                       #vsknn:  
                       lambda_ipw=3.62, 
                       lambda_idf=1
                       )

model_Ret.fit(train=train_data, test=test_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5),
    HitRate(10),
    HitRate(15),
    HitRate(20)]

output = evaluate_sessions(pr=model_Ret,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='VSTAN (Baseline)', reward_func='no reward func')
