==> 2.2 Diginetica Evaluation Trained Agent.ipynb <==
import pandas as pd
from src.recsys_baseline import VSKNN_STAN
from src.recsys_rl import rl_recommender
from src.evaluation import evaluate_sessions, MRR, HitRate, StrategicHybrid
from src.utils import evaluation_results_to_csv

## load datasets

train = pd.read_csv('data/processed datasets/diginetica/interactions_train_tr.txt',
                    sep='\t')
display(train.head())
display(train.shape)

test = pd.read_csv('data/processed datasets/diginetica/interactions_test.txt',
                    sep='\t', nrows=200)
display(test.head())
display(test.shape)

## Hybrid model  (reward: click=3, purchase=10)

### 5 interactions

model_1 = rl_recommender(
    n_history=5,
    file_path=r'trained agents\Diginetica\click=3, purchase=10\20240609_101255 - n_hist 5 state_size 5 action_size 568638.pth',
    state_size=5,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[5]) # after 5 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=5', reward_func='click=3, purchase=10')

### 10 interactions

model_1 = rl_recommender(
    n_history=10,
    file_path=r'trained agents\Diginetica\click=3, purchase=10\20240609_122056 - n_hist 10 state_size 10 action_size 568638.pth',
    state_size=10,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[10]) # after 10 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=10', reward_func='click=3, purchase=10')

### 15 interactions

model_1 = rl_recommender(
    n_history=15,
    file_path=r'trained agents\Diginetica\click=3, purchase=10\20240609_123637 - n_hist 15 state_size 15 action_size 568638.pth',
    state_size=15,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[15]) # after 15 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=15', reward_func='click=3, purchase=10')

### 100 interactions
In other words, the model almost only uses the first model to predict next interactions with items.

model_1 = rl_recommender(
    n_history=100,
    file_path=r'trained agents\Diginetica\click=3, purchase=10\20240609_134910 - n_hist 100 state_size 100 action_size 568638.pth',
    state_size=100,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[100]) # after 100 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=100', reward_func='click=3, purchase=10')

## Hybrid model (reward: click=0, purchase=10)

### 5 interactions

model_1 = rl_recommender(
    n_history=5,
    file_path=r'trained agents\Diginetica\click=0, purchase=10\20240627_190056 - n_hist 5 state_size 5 action_size 568638.pth',
    state_size=5,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[5]) # after 5 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=5', reward_func='click=0, purchase=10')

### 10 interactions

model_1 = rl_recommender(
    n_history=10,
    file_path=r'trained agents\Diginetica\click=0, purchase=10\20240627_190955 - n_hist 10 state_size 10 action_size 568638.pth',
    state_size=10,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[10]) # after 10 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=10', reward_func='click=0, purchase=10')

### 15 interactions

model_1 = rl_recommender(
    n_history=15,
    file_path=r'trained agents\Diginetica\click=0, purchase=10\20240627_191855 - n_hist 15 state_size 15 action_size 568638.pth',
    state_size=15,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[15]) # after 15 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=15', reward_func='click=0, purchase=10')

### 100 interactions
In other words, the model almost only uses the first model to predict next interactions with items.

model_1 = rl_recommender(
    n_history=100,
    file_path=r'trained agents\Diginetica\click=0, purchase=10\20240627_192759 - n_hist 100 state_size 100 action_size 568638.pth',
    state_size=100,
    action_size=568638,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=100,
                     sample_size=1000,
                     similarity='vec',
                    #    stan:  
                     lambda_spw=4.9,  
                     lambda_snh=80,  
                     lambda_inh=9.8,  
                    #    vsknn:  
                     lambda_ipw=4.9, 
                     lambda_idf=5
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[100]) # after 100 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]


output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test,
                  train_data=train,
                  cut_off=20)
evaluation_results_to_csv(output=output, dataset_name='Diginetica', algo_name='RL Agent n=100', reward_func='click=0, purchase=10')
