==> 2.2 Retailrocket Evaluation Trained Agent.ipynb <==
import pandas as pd
from src.recsys_baseline import VSKNN_STAN
from src.recsys_rl import rl_recommender
from src.evaluation import evaluate_sessions, MRR, HitRate, StrategicHybrid
from src.utils import evaluation_results_to_csv

## load datasets

train_data = pd.read_csv(r'data\processed datasets\retailrocket\events_train.csv',
                    sep='\t',)
display(train_data.head())
display(train_data.shape)

test_data = pd.read_csv(r'data\processed datasets\retailrocket\events_test.csv',
                    sep='\t',)

display(test_data.head())
display(test_data.shape)


## Hybrid model (reward: click=3, addtocart=8, purchase=10)

### 5 interactions

model_1 = rl_recommender(
    n_history=5,
    file_path=r'trained agents/retailrocket/20240704_083204 - n_hist 5 state_size 5 action_size 466865.pth',
    state_size=5,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[5]) # after 5 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=5', reward_func='click=3, addtocart=8, purchase=10')

### 10 interactions

model_1 = rl_recommender(
    n_history=10,
    file_path=r'trained agents/retailrocket/20240704_084626 - n_hist 10 state_size 10 action_size 466865.pth',
    state_size=10,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[10]) # after 10 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=10', reward_func='click=3, addtocart=8, purchase=10')

### 15 interactions

model_1 = rl_recommender(
    n_history=15,
    file_path=r'trained agents/retailrocket/20240704_083204 - n_hist 5 state_size 5 action_size 466865.pth',
    state_size=15,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[15]) # after 10 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=15', reward_func='click=3, addtocart=8, purchase=10')

### 100 interactions
In other words, the model almost only uses the first model to predict next interactions with items.

model_1 = rl_recommender(
    n_history=100,
    file_path=r'trained agents/retailrocket/20240704_083204 - n_hist 5 state_size 5 action_size 466865.pth',
    state_size=100,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[100]) # after 100 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=100', reward_func='click=3, addtocart=8, purchase=10')

## Hybrid model (reward: click=0, addtocart=8, purchase=10)

### 5 interactions

model_1 = rl_recommender(
    n_history=5,
    file_path=r'trained agents/Retailrocket/click=0, addtocart=8, purchase=10/20240704_100319 - n_hist 5 state_size 5 action_size 466865.pth',
    state_size=5,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[5]) # after 5 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=5', reward_func='click=0, addtocart=8, purchase=10')

### 10 interactions

model_1 = rl_recommender(
    n_history=10,
    file_path=r'trained agents/Retailrocket/click=0, addtocart=8, purchase=10/20240704_101830 - n_hist 10 state_size 10 action_size 466865.pth',
    state_size=10,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[10]) # after 10 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=10', reward_func='click=0, addtocart=8, purchase=10')

### 15 interactions

model_1 = rl_recommender(
    n_history=15,
    file_path=r'trained agents/Retailrocket/click=0, addtocart=8, purchase=10/20240704_110219 - n_hist 15 state_size 15 action_size 466865.pth',
    state_size=15,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[15]) # after 15 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=15', reward_func='click=0, addtocart=8, purchase=10')

### 100 interactions
In other words, the model almost only uses the first model to predict next interactions with items.

model_1 = rl_recommender(
    n_history=100,
    file_path=r'',
    state_size=100,
    action_size=466865,
    mode='predicting'
)

model_2 = VSKNN_STAN(k=500,
                     sample_size=1000,
                     similarity='cosine',
                    #    stan:  
                     lambda_spw=7.24,  
                     lambda_snh=100,  
                     lambda_inh=3.62,  
                    #    vsknn:  
                     lambda_ipw=3.62, 
                     lambda_idf=1
                       )

# Initialize the hybrid model
model = StrategicHybrid(algorithms=[model_1, model_2],
                        thresholds=[100]) # after 15 recommendations, the second algorithm is used
# Fit the hybrid model
model.fit(train_data)

metrics = [
    MRR(2),
    MRR(3),
    MRR(4),
    MRR(5),
    MRR(10),
    MRR(15),
    MRR(20),
    HitRate(1),
    HitRate(2),
    HitRate(3),
    HitRate(4),
    HitRate(5)]

output = evaluate_sessions(pr=model,
                  metrics= metrics,
                  test_data=test_data,
                  train_data=train_data,
                  cut_off=20)

evaluation_results_to_csv(output=output, dataset_name='Retailrocket', algo_name='RL Agent n=100', reward_func='click=0, addtocart=8, purchase=10')
